allow building with docker instead of podman

+++ compile-native.sh
@@ -38,8 +38,17 @@ echo "Compiling for: ${classifier} (dockcross: ${dockcross_image})"
 compiler_output_dir="${compiler_dir}/${classifier}"
 mkdir -p "$compiler_output_dir" 2>/dev/null
 proxy="${compiler_output_dir}/proxy"
-#podman pull "$dockcross_image"
-podman run --rm "$dockcross_image" >"$proxy"
+
+
+if [ -n "$(which podman)" ]
+then
+    podman run --rm "$dockcross_image" > "$proxy"
+elif [ -n "$(which docker)" ]
+then
+    docker run --rm "$dockcross_image" > "$proxy"
+else
+    echo "Either podman or docker are required to build JavaCAN!"
+fi
 chmod +x "$proxy"
 
 linker_output_dir="${relative_output_dir}/native/${classifier}/native"
